Android Application Based Home Automation System
a simple implementation of home automation system using Android Phone to enable the physically challenged or the elderly to switch the appliances with the help of smart phones. I specially learned Java for the development of the app. Embedded C coding was done in ATMEGA32 Microcontroller according to the algorithm. In the course of our work, we found that the dimmer circuit produced a flickering effect in the appliances instead of gradually dimming. To increase or decrease the intensity we created levels to the intensity from 1 to 5 - when total sine wave was received by the appliance the intensity would be maximum. So, interrupts were generated at the microcontroller whenever the level was increased or decreased through the android app. And according to the level the AC Sine wave was trimmed. In this way intensity variation was implemented and the problem was overcome.